Implement `add` command to assign tags to directories
The tags can be specified as command-line parameters, directories should
be listed in the prompt, or piped to stdin.

  $ shdo add "pictures" "personal"
  shdo> /home/myself/photos
  shdo> /home/mom/my_little_family

Tagging Git repositories under the current directory with
"git-repos" tag:

  $ find "`pwd`" -name '.git' -printf '%h\n' | sort -u \
  | shdo add "git-repos"

It's using `conf` package to store paths in user config dir.

[NOTE]
====
Piping to stding may not work on Git Bash, Bash says:

  stdin is not a tty

In this case try it again in cmd.exe or in PowerShell.
====

WARNING: Commit introducing TODOs