Separate view layer, handle `throw`
On the one hand, it is a good habit to separate the display from the
logic, on the other hand, that the GUI may later be available for the
program, so I have to use the display in an abstract way.

It really should have been a separate commit, but I couldn't determine
exactly the responsibilies: exceptions that are thrown while executing
commands (unknown shell command for example) are also handled by the
program and the user is warned about them.