Store most recently used directories and use as default
This is useful if the output is long and you want to execute additional
commands on certain directories. In this case, open a new terminal
window (so you can look back at results in the current terminal) and run
the program without arguments: the directory list is always stored when
tag or directory arguments are given, but if you run it without
arguments, it executes the commands on the last specified directories.